+++ zict/__init__.py
@@ -1,4 +1,5 @@
 from zict.buffer import Buffer
+from zict.cache import Cache, WeakValueMapping
 from zict.file import File
 from zict.func import Func
 from zict.lmdb import LMDB
@@ -6,4 +7,5 @@ from zict.lru import LRU
 from zict.sieve import Sieve
 from zict.zip import Zip
 
-__version__ = "2.2.0.dev1"
+# Must be kept aligned with setup.cfg
+__version__ = "2.2.0.dev2"